passes: Merge transformer with rewriter
